+++ server/src/routes/admin/index.ts
@@ -7,3 +7,4 @@ export { knowledgeAdminRoutes } from "./knowledge";
 export { personaAdminRoutes } from "./persona";
 export { pollAdminRoutes } from "./poll";
 export { userAdminRoutes } from "./users";
+export { widgetSiteAdminRoutes } from "./widget-sites";